ref: split and rename sentry.models.tests so it actually runs
