+++ tests/sentry/models/test_group.py
@@ -317,3 +317,13 @@ class GroupTest(TestCase, SnubaTestCase):
         assert group.get_last_release() == "100"
 
         assert group2.get_last_release() is None
+
+
+class GroupIsOverResolveAgeTest(TestCase):
+    def test_simple(self):
+        group = self.group
+        group.last_seen = timezone.now() - timedelta(hours=2)
+        group.project.update_option("sentry:resolve_age", 1)  # 1 hour
+        assert group.is_over_resolve_age() is True
+        group.last_seen = timezone.now()
+        assert group.is_over_resolve_age() is False
